Add LR scheduler and Fix Granite Checkpoint Loading (#80)
This PR adds the LR scheduler configuration and also fixes the granite loading issue for multinode settings.

this PR subsumes the changes in @aldo-pareja 's lr_scheduler branch, so a seperate merge of that would not be necessary
the strategy now for converting the granite checkpoint, is that each node will elect one local rank to do the convert.
the conversion will be one on the args.output_dir. There is no assumption it is the same for all nodes or not. But assuming they might be the same, the tmpdir is shared by a .{group_rank} postfix. See group rank
tested on multi node

---------

Signed-off-by: aldo pareja-cardona <aldo.pareja@ibm.com>
Signed-off-by: Yu Chin Fabian Lim <flim@sg.ibm.com>
Co-authored-by: aldo pareja-cardona <aldo.pareja@ibm.com>