Improve run_training error handling
The library shouldn't (just) print errors; it should raise errors and
let callers deal with them.

This patch also makes the function raise RuntimeError when training
process returns a non-zero return code, or when it times out on waiting
for exit.

ilab cli already handles all Exceptions raised by the function.
KeyboardInterrupt won't be caught by this exception handler since it's
not Exception (but BaseException), but it's handled by click library
instead [1].

The intent of this patch is to make `ilab train` fail with non-zero
return code when training routine failed.

[1] https://click.palletsprojects.com/en/7.x/exceptions/#where-are-errors-handled

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>