fix:  add fallback logic to use the distributed sampler
When we use the multipack sampler, it requires a certain shape of the dataset relative to the
GPUs to be able to sufficiently distribute all of the samples across different nodes.
When this happens, the train loaderbecomes empty which prevents us from being able to train.

This commit resolves that issue by falling back to the distributed sampler when the multipack
fails.

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>