fix: improve linting and automation
In the current version of the training repo, we have a few commands that help us lint
and check to see if there are any issues with types or incorrect imports. However;
there is no easy fix available when these issues are detected.

This commit introduces a few new functionalities:

1. Makefile receives 5 new commands:
 - toml-lint, which lints the pyproject.toml file
 - toml-fmt, which automatically formats the pyproject.toml file
 - check-engine, which checks that the specified container engine ($CENGINE) is installed
 - tests, which runs tests against the source repo
 - fix, which fixes all fixable issues

2. The tox.ini file receives a few additions:
  a. The testenv environment is defined with some preset values. Other targets are able to pull from this
  b. tox -e fix is added - automatically fixes all fixable problems
  c. tox -e tomllint is added - fixes all issues in the pyproject.toml file
  d. a [py] environment is added to specify the base python version

Aside from this, the resulting pyproject.toml file is added after formatting  is applied.

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>