feat: add total_samples as a field to logs being emitted
Today we emit a number of important values in the logs, but we do not explicitly announce how many samples
we are planning on iterating through. This commit emits total_samples as a field in both the initial log
as well as the subsequent logs during training. This way, this value can be accessed consistently by
any program interested in displaying information about its progress for convenience.

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>