fix: model.forward now accepts return_dict via kwargs
Since https://github.com/huggingface/transformers/pull/36794 it handles
the return_dict with can_return_tuple decorator that expects the
argument to be passed as kwarg.

Without the patch, instructlab functional tests fail with:

AttributeError: 'bool' object has no attribute 'unsqueeze'

This is because the passed return_dict argument is incorrectly
interpreted as cache_position.

Of course, the way the model forward method is overridden here is quite
problematic and would benefit from a refactor. This patch is a stop-gap
to fix the CI with minimal changes.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>