ci: Use tox-current-env to reuse prepared venv with torch
This is a completion of a workaround for flash-attn. The initial attempt
didn't work because we still used tox to run smoke tests, which means
another venv was created that did not include flash-attn.

To achieve sharing the venv for both torch installation and smoke tests,
we are using tox-current-env plugin here.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>