feat: Rely on implicit detection of distributed backend
From the official docs,

```
Since 2.6, if backend is not provided, c10d will use a backend
registered for the device type indicated by the device_id kwarg (if
provided).
```

and:

```
If neither backend nor device_id is provided, c10d will detect the
accelerator on the run-time machine and use a backend registered for
that detected accelerator (or cpu).
```

While the library is still cuda centric, this is one tiny step towards
a more agnostic implementation.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>