ci: add 3.12 smoke workflow flavor
Yes, it's a lot of duplication. Until we have some way to generate
workflows from a template, we have to have it: matrix doesn't apply
because if used, it will use the same single ec2 runner for both 3.11
and 3.12 runs - and while it works, it slows down feedback almost by x2
because these are run sequentially.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>