add and document building flash-attn with ninja
building flash-attn wheel was taking >2hr without ninja and higher
parallelism. this change adds ninja as a default requirement for the
package (because it needs to be installed _for_ flash-attn and is tiny).

this change also adds a short note about setting build parallelism with
the `MAX_JOBS` environment variable to configure the ninja build for
flash-attn

Signed-off-by: James Kunstle <jkunstle@redhat.com>