ci: Constrain all dependencies
Eventually we should adopt (some of this code) from ci-actions that
should host reusable parts of the constraints infra. The reusable
doesn't exist at the moment and is being cooked, so in the meantime,
it's ok to constrain it per repo.

(It also allows us to validate some CUDA specific quirks around
flash-attn before we commit to the shape of the reusable workflow
action.)

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>