adds build step for flash-attn to tox py3-smoke env
flash-attn can't be built inline with other dependencies because it
requires torch and ninja. we're currently getting around this with a
`tox_current_env` setup but that's tricky from a tox perspective. this
change makes it so that we can use tox to build the environment
directly, and that the environment won't be rebuilt in between command
runs.

Signed-off-by: James Kunstle <jkunstle@redhat.com>